add delete post control and mutation

+++ src/mutations/DeletePostMutation.js
@@ -1,9 +1,6 @@
-import {
-  commitMutation,
-  graphql,
-} from 'react-relay'
-import {ConnectionHandler} from 'relay-runtime'
-import environment from '../createRelayEnvironment'
+import { commitMutation, graphql } from 'react-relay'
+import { ConnectionHandler } from 'relay-runtime'
+import environment from '../Environment'
 
 const mutation = graphql`
   mutation DeletePostMutation($input: DeletePostInput!) {
@@ -11,28 +8,32 @@ const mutation = graphql`
       deletedId
     }
   }
-`;
+`
 
-export default function DeletePostMutation(postId) {
+export default function DeletePostMutation(postId, callback) {
   const variables = {
     input: {
       id: postId,
-      clientMutationId: ""
-    },
+      clientMutationId: ''
+    }
   }
-  commitMutation(
-    environment,
-    {
-      mutation,
-      variables,
-      onError: err => console.error(err),
-      updater: (proxyStore) => {
-        const deletePostField = proxyStore.getRootField('deletePost')
-        const deletedId = deletePostField.getValue('deletedId')
-        const viewerProxy = proxyStore.getRoot().getLinkedRecord('viewer')
-        const connection = ConnectionHandler.getConnection(viewerProxy, 'ListPage_allPosts')
+  commitMutation(environment, {
+    mutation,
+    variables,
+    onError: err => console.error(err),
+    onCompleted: response => callback && callback(),
+
+    updater: proxyStore => {
+      const deletePostField = proxyStore.getRootField('deletePost')
+      const deletedId = deletePostField.getValue('deletedId')
+      const viewerProxy = proxyStore.getRoot().getLinkedRecord('viewer')
+      const connection = ConnectionHandler.getConnection(
+        viewerProxy,
+        'PostList_allPosts'
+      )
+      if (connection) {
         ConnectionHandler.deleteNode(connection, deletedId)
       }
-    },
-  )
+    }
+  })
 }
